+++ src/dk/jonaslindstrom/mosef/modules/tuning/interpolate/LinearInterpolateFunction.java
@@ -3,7 +3,7 @@ package dk.jonaslindstrom.mosef.modules.tuning.interpolate;
 public class LinearInterpolateFunction implements InterpolateFunction {
 
   @Override
-  public float interpolate(float t) {
+  public double interpolate(double t) {
     return t;
   }
