+++ src/dk/jonaslindstrom/mosef/modules/tuning/interpolate/QuadraticInterpolateFunction.java
@@ -3,7 +3,7 @@ package dk.jonaslindstrom.mosef.modules.tuning.interpolate;
 public class QuadraticInterpolateFunction implements InterpolateFunction {
 
   @Override
-  public float interpolate(float t) {
+  public double interpolate(double t) {
     return t * t;
   }
